[Fix] my-page에서 표현하고자 하는 컴포너트를 분리
복용약 담당 컴포넌트와 관심 약국 담당 컴포넌트 분리
